Fix for [#ANDROID-129].
This involves handling contexts in a sane way. No more saving a static context in Apptentive and accessing it to save code verbosity. Same for resources such as SharedPreferences or Databases. All context access shall be through API methods that the host app calls, or checked for null before access, when needed (payload send thread).