AND-172 Add Who Card persistency check:
A boolean shared preference is introduced to track if the Who-Card is ever sent.  If email is required, this oolean will only be set true if the email address is valid and sent; If email is not required, no email validation will be imposed, and both skip and send actions will set the boolean to be true.

What's more:
1. If Who Card is being edited, rotating device will save and restore the states.
2. If Who Card is sent, the name/email/avatar file id will be saved to sharedpreference. Subsequent editing will prepopulate the values from the saved ones.